utils consolidaiton
